[refactor] move app state redux slices to appState
