+++ packages/client/src/editor.tsx
@@ -7,7 +7,7 @@ import "codemirror/mode/markdown/markdown.js";
 import "codemirror/lib/codemirror.css";
 import { WebsocketProvider } from "y-websocket";
 import { useSelector } from "react-redux";
-import { selectDocument } from "./redux/document";
+import { selectDocument } from "./redux/appState/document";
 
 const SIGNALLING_SERVICE =
   process.env.STORYBOOK_SIGNAL_URL || "ws://localhost:6006/ws/signal";
