+++ packages/client/src/redux/appState/user.ts
@@ -1,5 +1,5 @@
 import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import type { RootState } from "./store";
+import type { RootState } from "../store";
 
 interface UserState {
   me?: { id: string; username: string };
